RAG/requirements.txt
google-generativeai
pypdf
python-dotenv
fastapi
uvicorn
pydantic
pgvector
psycopg2-binary
supabase
requests
sentence-transformers
torch
